base: Declare a gradient end color
Our shishu themes uses a gradient for the bars, so we declared this color for using a more cool look on the gradients.
Going from an amber to our shishu purple is just disgusting, so why not doing it better?

A12 edit: Let's make it dynamic and use accent3 to complement accent1